src/requirements.txt
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
azure-core==1.29.7
azure-storage-queue==12.9.0
beautifulsoup4==4.12.3
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
cryptography==41.0.7
cymem==2.0.8
dataclasses-json==0.6.3
datasets==2.16.1
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docx2txt==0.8
elastic-transport==8.12.0
elasticsearch==8.12.0
evaluate==0.4.1
fastapi==0.109.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.10.0
gptcache==0.1.43
greenlet==3.0.3
guidance==0.0.64
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
interegular==0.3.3
isodate==0.6.1
Jinja2==3.1.3
joblib==1.3.2
jsonpath-ng==1.6.1
langcodes==3.3.0
llama-index==0.9.33
lm-format-enforcer==0.4.3
loguru==0.7.2
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mpmath==1.3.0
msal==1.26.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
onnx==1.15.0
onnxruntime==1.16.3
openai==1.8.0
optimum==1.16.1
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
preshed==3.0.9
protobuf==4.25.2
psutil==5.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.0.0
PyPDF2==3.0.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
rank-bm25==0.2.2
redis==5.0.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.11.4
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
wasabi==1.1.2
weasel==0.3.4
win32-setctime==1.1.0
wrapt==1.16.0
XlsxWriter==3.1.9
xxhash==3.4.1
yarl==1.9.4
